Reorganize lca modules
