+++ lib/main/tree/lca/main-lca-binary-lifting.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 #include "template-solver-interp.hpp"
-#include "cpalgo/tree/lowest-common-ancestor/lca-binary-lifting.hpp"
+#include "cpalgo/tree/lca/lca_binary_lifting.hpp"
 
 using namespace std;
