+++ lib/main/tree/lca/main-lca-farach-colton-bender.cpp
@@ -1,10 +1,10 @@
 #include <iostream>
 #include "template-solver-interp.hpp"
-#include "cpalgo/tree/lowest-common-ancestor/lca-segment-tree.hpp"
+#include "cpalgo/tree/lca/lca_farach_colton_bender.hpp"
 
 using namespace std;
 
-using Interp = SolverInterp<LCASegmentTree>;
+using Interp = SolverInterp<LCAFCB>;
 Interp* interp = new Interp();
 
 void setup(string& header, map<string,Command>& commands) {
