+++ lib/main/tree/lca/main-lca-segment-tree.cpp
@@ -1,10 +1,10 @@
 #include <iostream>
 #include "template-solver-interp.hpp"
-#include "cpalgo/tree/lowest-common-ancestor/lca-farach-colton-bender.hpp"
+#include "cpalgo/tree/lca/lca_segment_tree.hpp"
 
 using namespace std;
 
-using Interp = SolverInterp<LCAFCB>;
+using Interp = SolverInterp<LCASegmentTree>;
 Interp* interp = new Interp();
 
 void setup(string& header, map<string,Command>& commands) {
